requirements.txt
beautifulsoup4==4.9.3
apache-airflow-providers-mongo>=1.0.1